+++ src/main/java/com/fastcampus/springpractice/Main.java
@@ -1,0 +1,17 @@
+package com.fastcampus.springpractice;
+
+import com.fastcampus.springpractice.config.Config;
+import com.fastcampus.springpractice.service.SortService;
+import org.springframework.context.ApplicationContext;
+import org.springframework.context.annotation.AnnotationConfigApplicationContext;
+
+import java.util.Arrays;
+
+public class Main {
+    public static void main(String[] args) {
+        ApplicationContext context = new AnnotationConfigApplicationContext(Config.class);
+        SortService sortService = context.getBean(SortService.class);
+
+        System.out.println("[result] " + sortService.doSort(Arrays.asList(args)));
+    }
+}
