+++ src/main/java/com/fastcampus/springpractice/logic/BubbleSort.java
@@ -1,8 +1,11 @@
-package com.fastcampus.javaoop.logic;
+package com.fastcampus.springpractice.logic;
+
+import org.springframework.stereotype.Component;
 
 import java.util.ArrayList;
 import java.util.List;
 
+@Component
 public class BubbleSort<T extends Comparable<T>> implements Sort<T> {
 
     @Override
