+++ src/main/java/com/fastcampus/springpractice/service/SortService.java
@@ -1,0 +1,21 @@
+package com.fastcampus.springpractice.service;
+
+import com.fastcampus.springpractice.logic.Sort;
+import org.springframework.beans.factory.annotation.Qualifier;
+import org.springframework.stereotype.Service;
+
+import java.util.List;
+
+@Service
+public class SortService {
+
+    private Sort<String> sort;
+
+    public SortService(@Qualifier("bubbleSort") Sort<String> sort) {
+        this.sort = sort;
+    }
+
+    public List<String> doSort(List<String> list) {
+        return sort.sort(list);
+    }
+}
