+++ src/test/java/com/fastcampus/springpractice/service/SortServiceTest.java
@@ -1,6 +1,6 @@
-package com.fastcampus.javaoop.service;
+package com.fastcampus.springpractice.service;
 
-import com.fastcampus.javaoop.logic.JavaSort;
+import com.fastcampus.springpractice.logic.JavaSort;
 import org.junit.jupiter.api.Test;
 
 import java.util.List;
